chore: move utility func to file
